// src/startup/bootstrap/enums/muspell.enums.ts
/**
 * Enumeration with view engine options
 *
 * @enum EViewEngine
 */
export enum EViewEngine {
  PUG = 'pug',
  HANDLEBARS = 'hbs'
}

/**
 * Enumaration with yggdrasil application types
 *
 * @enum EApplicationType
 */
export enum EApplicationType {
  REST = 'REST',
  WEB = 'WEB',
  HYBRID = 'HYBRID'
}